feat(aeiou): add support for span information
This change adds support for reading new span information in the
`Layer`. The name of the span, its `span::Id` and fields are all
collected and stored. When an event is printed, it now includes all of
the spans in that event's hierarchy as well.

No span events are printed.

There are 3 examples, one showing only events, another showing events
with span parents, and a third that runs a small tokio task which will
be output if Tokio's tracing is enabled.

The README has also be updated to show how to use the examples.
Including the use of the `tokio_unstable` flag for the `tokio-tasks`
example.