feat(aeiou): Colorize output
The primary goal of `aeiou` is to provide special output for the
tracingt instrumentation present in Tokio. One way of providing that is
to colorize those traces specifically. This will make them easier to
distinguish amongst other traces, and should also make that content
easier to separate within itself. For instance to quickly tell whether a
line is for a task or a resource.

Use the `colored` crate to colorize the output of the subscriber layer.
The date and the level are colored.

More importantly, the known spans and events from Tokio's tracing
instrumentation is all explicitly colored.