feat(aeiou): add basic span events
It is often useful to be able to see when spans change state, as well as
the spans that form the scope of an event.

This change adds basic span events. Unlike the `tracing_subscriber` fmt
layer, these span events don't record the timing of the spans. Span
events don't include the target of the span and there are currently no
additional fields, just the message which will be one of:
* `new`
* `enter`
* `exit`
* `close`

As the span event message is a short message, it has been styled both
**bold** and __underlined__ to aid visibility.