chore(ari-subscriber): rename crate to `ari-subscriber`
I believed that the crate named `aeiou` was available on crates.io. It
turns out I was wrong. It isn't available, and hasn't been for a number
of years.

As such, I needed to come up with a new name. The new name is
`ari-subscriber`, where `ari` is an acronym for "Async Runtime
Instrumentation".